Move Activity class to separate file
Because it is going to be used by two other classes now
